Add migrateMappings foundation
Doesn't migrate yet, it just creates a diff between the current mappings and the
target (hardcoded for now) and writes it to a file.

I had to add the Paper and Sonatype repos for each subproject so they wouldn't
complain about not being able to find Mercury. I don't know if there's a better
to solve this.